Activity 2 Queries
List all city information for cities with a population greater than 1 million.
List the country codes of countries that have cities with a population greater than 1 million.
For each country, list the city (or cities) with the largest population. 
Select the country code of the country, followed by the province that the city is located in, along with the city name.
For each country, list its population and the total population of all neighboring countries. 
Select the country code of the country followed by its population, followed by the total population of all neighboring countries (and if it has no neighboring countries, NULL).
List countries having a border with another country.
List countries having at least two borders with another country.
List countries having no border with another country.
List all countries in Europe or Asia.
List all countries that are located in both Europe and Asia.
Find the city (or cities) with the largest population.